being refactor of parse; allow more types
objects in the namespace should be able to be of arbitrary types,
including types that require other objects to be passed in their init
params. To do that (previously init_params had to be something YAML
could immediately parse into an object) we need to know what type to
create and how we create that type.

The init_params can thus themselves define other objects which need to
first be instantiated before they can be passed in as params to the
containing node factory method

This starts down that road now.