Add HandleNode
Nodes returned from a NamespaceHandle need to override their NSIDs
If they don't, then you can't get a node from a NamespaceHandle and ever
use its NSID in the Handle that returned the node.

This changes the semantics of NamespaceHandle to only return HandleNodes
and HandleNodes wrap a real node and override the NSID to show the NSID
that the NamespaceHandle can use to address the Node.